feat(conect_api): criando contexto
