+++ src/pages/Transactions/index.tsx
@@ -1,31 +1,12 @@
-import { useEffect, useState } from 'react'
+import { useContext } from 'react'
 import { Header } from '../../components/Header'
 import { SearchForm } from '../../components/SearchForm'
 import { Summary } from '../../components/Summary'
 import { Transaction } from '../../components/Transaction'
-
-interface TransactionType {
-  id: number
-  description: string
-  expense: 'income' | 'outcome'
-  category: string
-  price: number
-  createdAt: string
-}
+import { TransactionsContext } from '../../contexts/TransactionContext'
 
 export const Transactions = () => {
-  const [transactions, setTransactions] = useState<TransactionType[]>([])
-
-  async function loadTransactions() {
-    const response: Response = await fetch('http://localhost:3000/transactions')
-    const data: TransactionType[] = await response.json()
-
-    setTransactions(data)
-  }
-
-  useEffect(() => {
-    loadTransactions()
-  }, [])
+  const { transactions } = useContext(TransactionsContext)
 
   return (
     <>
